fix: change to polish everything up

+++ 1stSemester/wip/lab3/zad1/palindrom.c
@@ -6,7 +6,7 @@
 #include "funs.h"
 
 bool palindrom(char napis[]) {
-    for (int i = 0; i < (strlen(napis))/2; i++) {
+    for (int i = 0; i < (int) (strlen(napis))/2; i++) {
         if(napis[i] != napis[strlen(napis) - 1 - i]) {
             return false;
         }
